Material Ui separate styles files and reusable compponents
