+++ src/App.js
@@ -1,6 +1,6 @@
 import { Route, Routes } from "react-router-dom";
 import HomePage from "./pages/HomePage";
-import Login from "./pages/Login";
+import Login from "./pages/Login/Login";
 import "./sass/main.scss";
 
 function App() {
