+++ src/pages/Login/LoginStyles.js
@@ -1,0 +1,45 @@
+const backgroundBlurImage = {
+  width: "100%",
+  height: "100vh",
+  filter: "blur(2rem)",
+};
+
+const welcomeStyleText = {
+  fontSize: "3.125rem",
+  fontWeight: 700,
+  letterSpacing: "0.5px",
+  lineHeight: "40px",
+  marginBottom: "1em",
+};
+
+const emailAddressText = {
+  position: "absolute",
+  top: ".9rem",
+  left: "1.45rem",
+  fontSize: "1.1rem",
+  fontWeight: 400,
+  lineHeight: 1.333,
+  color: "rgb(103, 112, 132)",
+};
+
+const newToContraText = {
+  color: "grey.dark",
+  maxWidth: "480px",
+  textAlign: "center",
+  fontWeight: "400",
+  fontSize: "1.7rem",
+  letterSpacing: ".15px",
+  lineHeight: " 1.42",
+};
+
+const loginImage = {
+  width: "90%",
+};
+
+export {
+  backgroundBlurImage,
+  welcomeStyleText,
+  emailAddressText,
+  newToContraText,
+  loginImage,
+};
